address comment, move jobmanager to server master
